[#7] NetworkService 구현
